refactor: move isGroupHandle into IMsgCore; extract ChatCache actor
